cleanliness 4/4: clear all lint debt + regression gate (bit-identical games)
Lint: repo is now fully clean (`ruff check src tests` passes). Fixed the 42
accumulated errors: unused imports (mostly in diagnostic harnesses), unsorted
import blocks, zip() strict= (strict=True where lengths are invariant-equal in
ai.py; strict=False documenting the deliberate truncation in naval's
landing-cell zip), an unnecessary lambda, a stray multiplication-sign glyph,
and test nits.

Regression gate for the whole 4-stage pass:
- suite 622 green at every stage;
- same-seed fingerprint proof: 4 seeds x 40 turns of portfolio-vs-baseline on
  the TUI's exact game builder, hashing full game state every turn
  (PYTHONHASHSEED=0), on this branch vs pre-cleanup main — ALL FOUR DIGESTS
  BIT-IDENTICAL. The cleanup changed zero behavior in the engine / generator /
  follower / naval / evaluator paths. (The one intended behavior change of the
  pass, the TUI unload-CAPTURE fix in 2/4, is TUI-only and outside this path.)

+++ src/empire/_eval_curves.py
@@ -13,10 +13,9 @@ Not a game; a probe of the value function. Run: python _eval_curves.py
 """
 from __future__ import annotations
 
-from empire._naval_arena import NAVAL_PROFILE, build_two_continent, _home_continent
+from empire._naval_arena import NAVAL_PROFILE, _home_continent, build_two_continent
 from empire.ai.search.evaluator import Evaluator, EvalWeights
 from empire.core.coord import Coord
-from empire.core.tile import TerrainKind
 from empire.core.unit import UNIT_REGISTRY, UnitKind
 
 # OLD = the pre-§10.2 surface (perimeter penalty, no opportunity/exploration).
@@ -59,7 +58,7 @@ def add_army(game, me, c):
 def bar(v, lo, hi, width=40):
     if hi <= lo:
         return ""
-    n = int(round((v - lo) / (hi - lo) * width))
+    n = round((v - lo) / (hi - lo) * width)
     return "#" * max(0, min(width, n))
